scanner/requirements.txt
aiosmtpd>=1.4
aiosmtplib>=1.4
yara-python>=4.2 
pyyaml>=6.0